requirements.txt
pyserial>=3.5
asciimatics>=1.14.0
